refactor: update entity
